[REFACTOR] move component from pages to components folder
